Mark Linux-specific logic and dependencies explicitly
