+++ ci/ios/test-router/raas/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "raas"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 description = "RaaS stands for Router as a Service"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -12,10 +12,8 @@ axum = "0.7"
 env_logger = "0.10"
 futures = "0.3"
 ipnetwork = { version = "0.21", features = ["serde"] }
+libc = "0.2"
 log = "0.4"
-mnl = { version = "0.3.1", features = ["mnl-1-0-4"] }
-nftnl = { version = "0.9", features = ["nftnl-1-1-0"] }
-once_cell = "1"
 pcap = { version = "1.2.0", features = ["capture-stream"] }
 pcap-file-tokio = "0.1"
 pnet_packet = "0.34"
@@ -26,6 +24,10 @@ tokio-util = { version = "0.7", features = ["io"] }
 tower = "0.4"
 tower-http = { version = "0.5", features = ["trace"] }
 uuid = { version = "1", features = ["serde"] }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+mnl = { version = "0.3.1", features = ["mnl-1-0-4"] }
+nftnl = { version = "0.9", features = ["nftnl-1-2-0"] }
 
 # This crate is excluded from the workspace, so it cannot inherit the lints.
 # Keep in sync with `[workspace.lints]` in the repository root `Cargo.toml`.
