+++ ci/ios/test-router/raas/src/firewall/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(target_os = "linux")]
+mod linux;
+
+mod rule;
+pub use rule::{BlockRule, Endpoints};
+
+#[cfg(target_os = "linux")]
+pub use linux::BlockList;
